requirements.txt
boto3==1.24.26
pandas==1.3.2
PyYAML==6.0
ipython==8.0.1